don't repeat operation

+++ setup.sh
@@ -5,8 +5,10 @@ PUB_KEY_FILE=$BASEDIR/*.pub
 SCREENRC_FILE=$BASEDIR/.screenrc
 ME=`whoami`
 
-cat $PUB_KEY_FILE >> ~/.ssh/authorized_keys*
-mv $SCREENRC_FILE ~/.screenrc
+if [ -e $SCREENRC_FILE ]; then
+    mv $SCREENRC_FILE ~/.screenrc    
+    cat $PUB_KEY_FILE >> ~/.ssh/authorized_keys*
+fi
 
 if [ $ME="root" ]; then
     yum -f install emacs screen tree git
